requirements.txt
requests
beautifulsoup4
ddgs
googlesearch-python